Separate import classes into files
